Move signal creation functions into the types module
The same code will also work for interfaces.
